refactor: Move codegen dependency out from zksync_types
