+++ core/lib/types/Cargo.toml
@@ -17,8 +17,6 @@ zksync_basic_types = { path = "../basic_types" }
 zksync_contracts = { path = "../contracts" }
 zksync_mini_merkle_tree = { path = "../mini_merkle_tree" }
 zksync_config = { path = "../config" }
-# We need this import because we wanat DAL to be responsible for (de)serialization
-codegen = { git = "https://github.com/matter-labs/solidity_plonk_verifier.git", branch = "dev" }
 zkevm_test_harness = { git = "https://github.com/matter-labs/era-zkevm_test_harness.git", branch = "v1.3.3" }
 zksync_protobuf = { version = "0.1.0", git = "https://github.com/matter-labs/era-consensus.git", rev = "5727a3e0b22470bb90092388f9125bcb366df613" }
