+++ core/lib/zksync_core/src/eth_sender/mod.rs
@@ -1,6 +1,7 @@
 mod aggregator;
 mod error;
 mod eth_tx_aggregator;
+mod eth_tx_args;
 mod eth_tx_manager;
 mod metrics;
 mod publish_criterion;
